买家商品列表

+++ src/main/java/com/kobe/wxorder/model/ProductInfo.java
@@ -2,8 +2,10 @@ package com.kobe.wxorder.model;
 
 import lombok.Data;
 import org.hibernate.annotations.DynamicUpdate;
+
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -22,7 +24,7 @@ public class ProductInfo {
     /** 商品名 */
     private String productName;
     /** 商品单价 */
-    private Double productPrice;
+    private BigDecimal productPrice;
     /** 商品库存 */
     private Integer productStock;
     /** 商品小图 */
